requirements/app.txt
fastapi
starlette
requests
mongoengine
PyJWT
pendulum
uvicorn
pymongo[srv]
loguru